feat: move app auth in auth create
